+++ CHANGELOG.rst
@@ -27,7 +27,7 @@ Changed
 - fixed issues with all null and nullable-bool column handling in dataframe_init_dispatch
 - added NaN error handling to WeightColumnMixin
 - narwhalified MeanImputer `#344 https://github.com/lvgig/tubular/issues/344_`
-- placeholder
+- narwhalified BaseNumericTransformer `#358 https://github.com/lvgig/tubular/issues/358`
 - placeholder
 - placeholder